Added function to play game up to winning score
Renamed the function that compares the computer an player selections to
playGame as it encompasses all of the important gameplay functions.
Created function to run game. Prints score after each round. Once
either the player or computer hits the winning score, runs a comparison
and logs the winner to the console.